runtime: streamline moduledata.textAddr
Accept a uint32 instead of a uintptr to make call sites simpler.

Do less work in the common case in which len(textsectmap) == 1.

Reviewed-on: https://go-review.googlesource.com/c/go/+/354132
Trust: Josh Bleecher Snyder <josharian@gmail.com>
Run-TryBot: Josh Bleecher Snyder <josharian@gmail.com>
TryBot-Result: Go Bot <gobot@golang.org>
(cherry picked from commit 8238f82bf1ddadd906585e3f11600968d95ce98c)

+++ src/runtime/type.go
@@ -285,7 +285,7 @@ func (t *_type) textOff(off textOff) unsafe.Pointer {
 		}
 		return res
 	}
-	res := md.textAddr(uintptr(off))
+	res := md.textAddr(uint32(off))
 	return unsafe.Pointer(res)
 }
